rename earv2 to earn
