+++ src/app/earn-old/components/chart-tabs.tsx
@@ -1,8 +1,9 @@
 import { Dispatch, SetStateAction } from 'react'
 
-import { ChartTab } from '@/app/earn/types'
 import { useAnalytics } from '@/lib/hooks/use-analytics'
 import { cn } from '@/lib/utils/tailwind'
+
+import { ChartTab } from '@/app/earn/types'
 
 type Tab = { name: ChartTab; label: string }
