+++ src/app/earn-old/components/earn-widget/styles.css
