+++ src/app/earn-old/components/faq-section/index.tsx
