+++ src/app/earn-old/provider.tsx
@@ -1,6 +1,7 @@
 'use client'
 
 import { getTokenByChainAndSymbol, isAddressEqual } from '@indexcoop/tokenlists'
+import { useQuery } from '@tanstack/react-query'
 import {
   createContext,
   useCallback,
@@ -8,22 +9,27 @@ import {
   useMemo,
   useState,
 } from 'react'
-import * as chains from 'viem/chains'
+import { isAddress } from 'viem'
+import { base } from 'viem/chains'
 
-import { useQueryParams } from '@/app/earn/use-query-params'
 import { ETH, type Token } from '@/constants/tokens'
-import { GetApiV2ProductsEarn200 } from '@/gen'
-import { type TokenBalance } from '@/lib/hooks/use-balance'
-import { useMultiChainBalances } from '@/lib/hooks/use-multichain-balances'
+import { type TokenBalance, useBalances } from '@/lib/hooks/use-balance'
 import { useNetwork } from '@/lib/hooks/use-network'
 import { useQuoteResult } from '@/lib/hooks/use-quote-result'
 import { useWallet } from '@/lib/hooks/use-wallet'
 import { useSlippage } from '@/lib/providers/slippage'
 import { isValidTokenInput, parseUnits } from '@/lib/utils'
+import {
+  fetchTokenHistoricalData,
+  fetchTokenMetrics,
+} from '@/lib/utils/api/index-data-provider'
+import { calculateApy } from '@/lib/utils/apy'
 
 import { getCurrencyTokens, getYieldTokens } from './constants'
 
 import type { QuoteResult } from '@/lib/hooks/use-best-quote/types'
+
+import { useQueryParams } from '@/app/earn/use-query-params'
 
 const hyEthTokenlist = getTokenByChainAndSymbol(1, 'hyETH')
 const hyETH = { ...hyEthTokenlist, image: hyEthTokenlist.logoURI }
@@ -35,13 +41,18 @@ interface Context {
   balances: TokenBalance[]
   indexToken: Token
   indexTokens: Token[]
-  products: GetApiV2ProductsEarn200
+  apy: number | null
+  apy7d: number | null
+  apy30d: number | null
+  nav: number | null
+  tvl: number | null
   inputToken: Token
   outputToken: Token
   inputTokenAmount: bigint
   inputTokens: Token[]
   outputTokens: Token[]
   isFetchingQuote: boolean
+  isFetchingStats: boolean
   quoteResult: QuoteResult | null
   onChangeInputTokenAmount: (input: string) => void
   onSelectIndexToken: (tokenSymbol: string, chainId: number) => void
@@ -57,13 +68,18 @@ export const EarnContext = createContext<Context>({
   balances: [],
   indexToken: hyETH,
   indexTokens: [],
-  products: [],
+  apy: null,
+  apy7d: null,
+  apy30d: null,
+  nav: null,
+  tvl: null,
   inputToken: ETH,
   outputToken: hyETH,
   inputTokenAmount: BigInt(0),
   inputTokens: [],
   outputTokens: [],
   isFetchingQuote: false,
+  isFetchingStats: true,
   quoteResult: null,
   onChangeInputTokenAmount: () => {},
   onSelectIndexToken: () => {},
@@ -81,10 +97,7 @@ const defaultParams = {
   outputToken: hyETH,
 }
 
-export function EarnProvider(props: {
-  children: any
-  products: GetApiV2ProductsEarn200
-}) {
+export function EarnProvider(props: { children: any }) {
   const { chainId: chainIdRaw } = useNetwork()
   const { address } = useWallet()
   const {
@@ -100,7 +113,7 @@ export function EarnProvider(props: {
   const outputToken = queryOutputToken
 
   const chainId = useMemo(() => {
-    return chainIdRaw ?? chains.base.id
+    return chainIdRaw ?? base.id
   }, [chainIdRaw])
 
   const indexToken = useMemo(() => {
@@ -108,11 +121,20 @@ export function EarnProvider(props: {
     return inputToken
   }, [inputToken, isMinting, outputToken])
 
+  const indexTokenAddress = indexToken.address ?? ''
+
   const indexTokens = useMemo(() => {
     return getYieldTokens()
   }, [])
 
-  const { data: balances } = useMultiChainBalances(props.products)
+  const indexTokenAddresses = useMemo(() => {
+    return indexTokens.map((token) => token.address!)
+  }, [indexTokens])
+
+  const { balances, forceRefetchBalances } = useBalances(
+    address,
+    indexTokenAddresses,
+  )
 
   const inputTokens = useMemo(() => {
     const isIcEth = isAddressEqual(indexToken.address, icETH.address)
@@ -143,6 +165,73 @@ export function EarnProvider(props: {
     inputTokenAmount,
     inputValue,
     slippage,
+  })
+
+  const {
+    data: { apy, nav, tvl },
+    isFetching: isFetchingLatestStats,
+  } = useQuery({
+    enabled: isAddress(indexTokenAddress),
+    gcTime: 2 * 60 * 1000,
+    refetchOnWindowFocus: false,
+    initialData: { apy: null, nav: null, tvl: null },
+    queryKey: ['token-latest-stats', indexTokenAddress],
+    queryFn: async () => {
+      const data = await fetchTokenMetrics({
+        chainId: indexToken.chainId!,
+        tokenAddress: indexTokenAddress!,
+        metrics: ['nav', 'apy', 'pav'],
+      })
+
+      return {
+        apy: data?.APY ? calculateApy(data) : null,
+        nav: data?.NetAssetValue ?? null,
+        tvl: data?.ProductAssetValue ?? null,
+      }
+    },
+  })
+
+  const {
+    data: { apy7d, apy30d },
+    isFetching: isFetchingApyStats,
+  } = useQuery({
+    enabled: isAddress(indexTokenAddress),
+    gcTime: 2 * 60 * 1000,
+    refetchOnWindowFocus: false,
+    initialData: { apy7d: null, apy30d: null },
+    queryKey: ['token-apy-stats', indexTokenAddress],
+    queryFn: async () => {
+      const data = await fetchTokenHistoricalData({
+        chainId,
+        tokenAddress: indexTokenAddress!,
+        metrics: ['apy'],
+        interval: 'daily',
+        period: 'month',
+      })
+
+      const historicalData = (data ?? []).sort(
+        (a, b) =>
+          new Date(b.CreatedTimestamp).getTime() -
+          new Date(a.CreatedTimestamp).getTime(),
+      )
+
+      const apy7d = historicalData
+        .slice(0, 7)
+        .reduce((acc, current, _, { length }) => {
+          return acc + calculateApy(current) / length
+        }, 0)
+
+      const apy30d = historicalData
+        .slice(0, 30)
+        .reduce((acc, current, _, { length }) => {
+          return acc + calculateApy(current) / length
+        }, 0)
+
+      return {
+        apy7d: apy7d ? apy7d : null,
+        apy30d: apy30d ? apy30d : null,
+      }
+    },
   })
 
   const onChangeInputTokenAmount = useCallback(
@@ -203,6 +292,7 @@ export function EarnProvider(props: {
   const reset = () => {
     setInputValue('')
     resetQuote()
+    forceRefetchBalances()
   }
 
   const toggleIsMinting = useCallback(() => {
@@ -225,10 +315,15 @@ export function EarnProvider(props: {
         inputToken,
         outputToken,
         inputTokenAmount,
-        products: props.products,
+        apy,
+        apy7d,
+        apy30d,
+        nav,
+        tvl,
         inputTokens,
         outputTokens,
         isFetchingQuote,
+        isFetchingStats: isFetchingLatestStats || isFetchingApyStats,
         quoteResult,
         onChangeInputTokenAmount,
         onSelectIndexToken,
