+++ src/app/earn-old/types.ts
