+++ src/app/earn/components/balance-card.tsx
