+++ src/app/earn/components/earn-widget/components/caption.tsx
