+++ src/app/earn/components/earn-widget/components/projection.tsx
