+++ src/app/earn/components/earn-widget/components/smart-trade-button.tsx
