+++ src/app/earn/components/earn-widget/components/trade-input-selector.tsx
