+++ src/app/earn/components/earn-widget/index.tsx
@@ -2,33 +2,28 @@
 
 import { useAtom } from 'jotai'
 import { useCallback, useEffect } from 'react'
+import { isAddressEqual } from 'viem'
 
-import { Summary } from '@/app/earn/components/earn-widget/components/summary'
 import { supportedNetworks } from '@/app/earn/constants'
 import { useEarnContext } from '@/app/earn/provider'
-import { useQueryParams } from '@/app/earn/use-query-params'
+import { useQueryParams } from '@/app/earn-old/use-query-params'
 import { tradeMachineAtom } from '@/app/store/trade-machine'
-import { Receive } from '@/components/receive'
-import { BuySellSelector } from '@/components/selectors/buy-sell-selector'
-import { Settings } from '@/components/settings'
-import { SmartTradeButton } from '@/components/smart-trade-button'
-import { SelectTokenModal } from '@/components/swap/components/select-token-modal'
-import { TradeInputSelector } from '@/components/swap/components/trade-input-selector'
 import { TransactionReviewModal } from '@/components/swap/components/transaction-review'
 import { WarningType } from '@/components/swap/components/warning'
 import { TradeButtonState } from '@/components/swap/hooks/use-trade-button-state'
-import { TokenDisplay } from '@/components/token-display'
 import { useDisclosure } from '@/lib/hooks/use-disclosure'
 import { useGasData } from '@/lib/hooks/use-gas-data'
 import { useSupportedNetworks } from '@/lib/hooks/use-network'
-import { useWallet } from '@/lib/hooks/use-wallet'
 import { useSlippage } from '@/lib/providers/slippage'
 import { formatWei } from '@/lib/utils'
 import { getMaxBalance } from '@/lib/utils/max-balance'
 
 import { useFormattedEarnData } from '../../use-formatted-data'
 
-import './styles.css'
+import { DepositWithdraw } from './components/deposit-withdraw'
+import { Projection } from './components/projection'
+import { SmartTradeButton } from './components/smart-trade-button'
+import { TradeInputSelector } from './components/trade-input-selector'
 
 const hiddenLeverageWarnings = [WarningType.flashbots]
 
@@ -36,18 +31,16 @@ export function EarnWidget() {
   const gasData = useGasData()
   const isSupportedNetwork = useSupportedNetworks(supportedNetworks)
   const { queryParams } = useQueryParams()
-  const { address } = useWallet()
+
   const {
     indexToken,
     inputToken,
     inputTokenAmount,
-    inputTokens,
     inputValue,
     isMinting,
-    outputTokens,
     onChangeInputTokenAmount,
-    onSelectInputToken,
-    onSelectOutputToken,
+    balances,
+    products,
     outputToken,
     reset,
     toggleIsMinting,
@@ -60,29 +53,23 @@ export function EarnWidget() {
     inputBalance,
     inputBalanceFormatted,
     isFetchingQuote,
-    ouputAmount,
     resetData,
   } = useFormattedEarnData()
 
-  const {
-    isOpen: isSelectInputTokenOpen,
-    onOpen: onOpenSelectInputToken,
-    onClose: onCloseSelectInputToken,
-  } = useDisclosure()
-  const {
-    isOpen: isSelectOutputTokenOpen,
-    onOpen: onOpenSelectOutputToken,
-    onClose: onCloseSelectOutputToken,
-  } = useDisclosure()
+  const selectedProduct = products.find((p) =>
+    isAddressEqual(p.tokenAddress, indexToken?.address ?? ''),
+  )
+
+  const { onOpen: onOpenSelectInputToken } = useDisclosure()
 
   const [tradeState, sendTradeEvent] = useAtom(tradeMachineAtom)
 
   const {
-    auto: autoSlippage,
-    isAuto: isAutoSlippage,
-    set: setSlippage,
+    // auto: autoSlippage,
+    // isAuto: isAutoSlippage,
+    // set: setSlippage,
     setSlippageForToken,
-    slippage,
+    // slippage,
   } = useSlippage()
 
   const onClickBalance = useCallback(() => {
@@ -104,22 +91,16 @@ export function EarnWidget() {
   }, [tradeState, reset, resetData, sendTradeEvent])
 
   return (
-    <div className='earn-widget flex h-fit flex-col gap-3 rounded-lg px-4 py-6 lg:ml-auto'>
-      <div className='flex justify-between'>
-        <TokenDisplay mini token={indexToken} />
-        <Settings
-          isAuto={isAutoSlippage}
-          isDarkMode={false}
-          slippage={slippage}
-          onChangeSlippage={setSlippage}
-          onClickAuto={autoSlippage}
-        />
-      </div>
-      <BuySellSelector isMinting={isMinting} onClick={toggleIsMinting} />
+    <div className='earn-widget flex h-fit flex-col gap-6'>
+      <DepositWithdraw
+        isMinting={isMinting}
+        toggleIsMinting={toggleIsMinting}
+      />
       <TradeInputSelector
+        showSelectorButtonChevron={isMinting}
         config={{ isReadOnly: false }}
         balance={inputBalanceFormatted}
-        caption='You pay'
+        caption={isMinting ? 'Deposit' : 'Withdraw'}
         formattedFiat={inputAmoutUsd}
         selectedToken={inputToken}
         selectedTokenAmount={inputValue}
@@ -127,13 +108,16 @@ export function EarnWidget() {
         onClickBalance={onClickBalance}
         onSelectToken={onOpenSelectInputToken}
       />
-      <Receive
-        isLoading={isFetchingQuote}
-        outputAmount={ouputAmount}
-        selectedOutputToken={outputToken}
-        onSelectToken={onOpenSelectOutputToken}
+      <Projection
+        amount={inputValue}
+        balance={
+          balances.find((b) => b.token === indexToken.address)?.value ??
+          BigInt(0)
+        }
+        product={selectedProduct}
+        isMinting={isMinting}
       />
-      <Summary />
+
       <SmartTradeButton
         contract={contract ?? ''}
         hasFetchingError={false}
@@ -152,29 +136,8 @@ export function EarnWidget() {
         onOpenTransactionReview={() => sendTradeEvent({ type: 'REVIEW' })}
         onRefetchQuote={() => {}}
       />
-      <SelectTokenModal
-        isOpen={isSelectInputTokenOpen}
-        onClose={onCloseSelectInputToken}
-        onSelectedToken={(tokenSymbol, chainId) => {
-          onSelectInputToken(tokenSymbol, chainId)
-          onCloseSelectInputToken()
-        }}
-        address={address}
-        tokens={inputTokens}
-        showNetworks={!isMinting}
-      />
-      <SelectTokenModal
-        isOpen={isSelectOutputTokenOpen}
-        onClose={onCloseSelectOutputToken}
-        onSelectedToken={(tokenSymbol, chainId) => {
-          onSelectOutputToken(tokenSymbol, chainId)
-          onCloseSelectOutputToken()
-        }}
-        address={address}
-        tokens={outputTokens}
-        showNetworks={isMinting}
-      />
       <TransactionReviewModal
+        isDarkMode
         onClose={() => {
           reset()
           resetData()
