+++ src/app/earn/components/product-pill.tsx
