+++ src/app/earn/components/product-tag.tsx
