+++ src/app/earn/components/stat-box.tsx
