+++ src/components/header/network-select.tsx
@@ -2,9 +2,9 @@
 
 import { Button, Popover, PopoverButton, PopoverPanel } from '@headlessui/react'
 import { ChevronDownIcon, XMarkIcon } from '@heroicons/react/20/solid'
+import { useAppKit } from '@reown/appkit/react'
 import { NetworkUtil } from '@reown/appkit-common'
 import { AssetUtil, ChainController } from '@reown/appkit-core'
-import { useAppKit } from '@reown/appkit/react'
 import { watchAccount } from '@wagmi/core'
 import { AnimatePresence, motion } from 'framer-motion'
 import Image from 'next/image'
